requirements.txt
fastapi[all]==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
requests==2.32.3
beautifulsoup4==4.13.4
openai==1.75.0
python-dotenv==1.1.0
groq==0.13.0